+++ src/features/backups/QuickSave.cpp
@@ -1,16 +1,28 @@
+#include "QuickSave.hpp"
 #include <Geode/modify/EditorPauseLayer.hpp>
 #include <Geode/modify/GManager.hpp>
 #include <Geode/modify/MenuLayer.hpp>
 #include <cvolton.level-id-api/include/EditorIDs.hpp>
 #include <hjfod.gmd-api/include/GMD.hpp>
+#include <utils/HolyUB.hpp>
 
 using namespace geode::prelude;
 
 static std::filesystem::path getQuickSaveDir() {
     return Mod::get()->getSaveDir() / "quicksave";
 }
-static std::filesystem::path getQuickSaveFile(GJGameLevel* level) {
-    return getQuickSaveDir() / fmt::format("{}.gmd", EditorIDs::getID(level));
+
+// Autosave is stored separately, as if you press Exit without saving in the 
+// editor then that should discard autosaves too, but not quicksaves
+static std::filesystem::path getAutoSaveDir() {
+    return Mod::get()->getSaveDir() / "autosave";
+}
+
+static bool CREATING_AUTO_SAVE = false;
+void ::createAutoSave(LevelEditorLayer* lel) {
+    CREATING_AUTO_SAVE = true;
+    fakeEditorPauseLayer(lel)->saveLevel();
+    CREATING_AUTO_SAVE = false;
 }
 
 static bool SKIP_NEXT_LLM_SAVE = false;
@@ -27,6 +39,7 @@ class $modify(GManager) {
         if (isLLM) {
             std::error_code ec;
             std::filesystem::remove_all(getQuickSaveDir(), ec);
+            std::filesystem::remove_all(getAutoSaveDir(), ec);
         }
 	}
 };
@@ -37,15 +50,49 @@ class $modify(EditorPauseLayer) {
         // In any case we need to still run the rest of `saveLevel` to have GD 
         // update the level state but for quicksave we just skip the function 
         // saving CCLocalLevels 
-        SKIP_NEXT_LLM_SAVE = Mod::get()->template getSettingValue<bool>("quick-save");
+        SKIP_NEXT_LLM_SAVE = Mod::get()->template getSettingValue<bool>("quick-save") || CREATING_AUTO_SAVE;
         EditorPauseLayer::saveLevel();
 
+        auto dir = CREATING_AUTO_SAVE ? getAutoSaveDir() : getQuickSaveDir();
+        CREATING_AUTO_SAVE = false;
+
         // Save to individual file regardless as a backup if saving crashes
-        (void)file::createDirectoryAll(getQuickSaveDir());
-        auto res = gmd::exportLevelAsGmd(m_editorLayer->m_level, getQuickSaveFile(m_editorLayer->m_level));
+        (void)file::createDirectoryAll(dir);
+        auto res = gmd::exportLevelAsGmd(
+            m_editorLayer->m_level,
+            dir / fmt::format("{}.gmd", EditorIDs::getID(m_editorLayer->m_level))
+        );
         if (!res) {
-            log::error("Unable to quicksave level '{}': {}", m_editorLayer->m_level->m_levelName, res.unwrapErr());
+            log::error("Unable to save level '{}': {}", m_editorLayer->m_level->m_levelName, res.unwrapErr());
         }
+    }
+
+    $override
+    void FLAlert_Clicked(FLAlertLayer* fl, bool btn2) {
+        // Discard autosaves on exit without save
+        if (fl->getTag() == 1 && btn2) {
+            log::warn("Discarding autosaved changes");
+            std::error_code ec;
+            std::filesystem::remove_all(
+                getAutoSaveDir() / fmt::format("{}.gmd", EditorIDs::getID(m_editorLayer->m_level)),
+                ec
+            );
+        }
+        EditorPauseLayer::FLAlert_Clicked(fl, btn2);
+    }
+
+    $override
+    void onExitNoSave(CCObject*) {
+        auto fl = FLAlertLayer::create(
+            this,
+            "Exit",
+            "<cy>Exit</c> without saving? All unsaved changes will be <cr>lost</c>!\n"
+            "<co>All auto-saved changes will also be discarded!</c>",
+            "Cancel", "Exit",
+            300
+        );
+        fl->setTag(1);
+        fl->show();
     }
 };
 class $modify(MenuLayer) {
@@ -54,7 +101,9 @@ class $modify(MenuLayer) {
             return false;
         
         // Check if there's quicksaved crash data
-        for (auto file : file::readDirectory(getQuickSaveDir()).unwrapOrDefault()) {
+        auto files = file::readDirectory(getQuickSaveDir()).unwrapOrDefault();
+        ranges::push(files, file::readDirectory(getAutoSaveDir()).unwrapOrDefault());
+        for (auto file : files) {
             auto data = gmd::importGmdAsLevel(file);
             if (!data) continue;
             auto imported = *data;