+++ src/features/backups/QuickSave.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <Geode/binding/GJGameLevel.hpp>
+
+using namespace geode::prelude;
+
+void createAutoSave(LevelEditorLayer* lel);